requirements.txt
requests
beautifulsoup4
PyYAML
pydantic